feat(vacancy): modularize workflows and ship demo api
- split the apollo importer, vacancy applications, and readiness reporting into focused modules with dedicated tests
- add the axum-based API/CLI harness plus demo tooling and infrastructure helpers for vacancy insights
- document the readiness narratives and include the patched Apollo dataset to anchor investor storylines
