+++ crates/tenant-ai/src/workflows/vacancy/applications/tests/mod.rs
@@ -1,0 +1,5 @@
+mod common;
+mod compliance;
+mod evaluation;
+mod routing;
+mod service;
