+++ crates/tenant-ai/src/workflows/vacancy/instance.rs
@@ -12,12 +12,12 @@ pub struct TaskDetailView {
     pub key: &'static str,
     pub name: &'static str,
     pub stage: VacancyStage,
-    pub stage_label: String,
+    pub stage_label: &'static str,
     pub role: VacancyRole,
-    pub role_label: String,
+    pub role_label: &'static str,
     pub due_date: NaiveDate,
     pub status: TaskStatus,
-    pub status_label: String,
+    pub status_label: &'static str,
     pub completed_on: Option<NaiveDate>,
     pub deliverables: Vec<&'static str>,
     pub compliance: Vec<ComplianceNote>,
@@ -42,12 +42,12 @@ impl TaskInstance {
             key: self.template.key,
             name: self.template.name,
             stage: self.template.stage,
-            stage_label: self.template.stage.label().to_string(),
+            stage_label: self.template.stage.label(),
             role: self.template.primary_role,
-            role_label: self.template.primary_role.label().to_string(),
+            role_label: self.template.primary_role.label(),
             due_date: self.due_date,
             status: self.status,
-            status_label: self.status.label().to_string(),
+            status_label: self.status.label(),
             completed_on: self.completed_on,
             deliverables: self.template.deliverables.clone(),
             compliance: self.template.compliance.clone(),
