+++ crates/tenant-ai/src/workflows/vacancy/report/mod.rs
@@ -1,0 +1,8 @@
+mod insights;
+mod summary;
+pub mod views;
+
+pub use summary::VacancyReport;
+
+pub(crate) use insights::generate_insights;
+pub(crate) use summary::{ComplianceAlert, TaskSnapshot};
