+++ crates/tenant-ai/src/workflows/vacancy/report/summary.rs
@@ -1,46 +1,11 @@
-use super::domain::{ComplianceSeverity, TaskStatus, VacancyRole, VacancyStage};
+use super::super::domain::{ComplianceSeverity, TaskStatus, VacancyRole, VacancyStage};
+use super::super::instance::VacancyWorkflowInstance;
+use super::views::{
+    ComplianceAlertView, RoleLoadEntry, StageProgressEntry, TaskSnapshotView, VacancyInsights,
+    VacancyReportSummary,
+};
 use chrono::NaiveDate;
-use serde::Serialize;
 use std::collections::HashMap;
-
-#[derive(Debug, Clone, Serialize)]
-pub struct StageProgressEntry {
-    pub stage: VacancyStage,
-    pub stage_label: String,
-    pub completed: usize,
-    pub total: usize,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct RoleLoadEntry {
-    pub role: VacancyRole,
-    pub role_label: String,
-    pub open: usize,
-    pub overdue: usize,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct TaskSnapshotView {
-    pub key: &'static str,
-    pub name: &'static str,
-    pub stage: VacancyStage,
-    pub stage_label: String,
-    pub role: VacancyRole,
-    pub role_label: String,
-    pub due_date: NaiveDate,
-    pub status: TaskStatus,
-    pub status_label: String,
-    pub completed_on: Option<NaiveDate>,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct ComplianceAlertView {
-    pub task_key: &'static str,
-    pub topic: &'static str,
-    pub detail: &'static str,
-    pub severity: ComplianceSeverity,
-    pub severity_label: String,
-}
 
 #[derive(Debug, Default, Clone)]
 pub struct StageProgress {
@@ -71,7 +36,7 @@ impl VacancyReport {
                     .get(&stage)
                     .map(|progress| StageProgressEntry {
                         stage,
-                        stage_label: stage.label().to_string(),
+                        stage_label: stage.label(),
                         completed: progress.completed,
                         total: progress.total,
                     })
@@ -83,7 +48,7 @@ impl VacancyReport {
             .filter_map(|role| {
                 self.role_load.get(&role).map(|load| RoleLoadEntry {
                     role,
-                    role_label: role.label().to_string(),
+                    role_label: role.label(),
                     open: load.open,
                     overdue: load.overdue,
                 })
@@ -111,12 +76,16 @@ impl VacancyReport {
     }
 }
 
-#[derive(Debug, Clone, Serialize)]
-pub struct VacancyReportSummary {
-    pub stage_progress: Vec<StageProgressEntry>,
-    pub role_load: Vec<RoleLoadEntry>,
-    pub overdue_tasks: Vec<TaskSnapshotView>,
-    pub compliance_alerts: Vec<ComplianceAlertView>,
+impl VacancyReportSummary {
+    pub fn insights(
+        &self,
+        instance: &VacancyWorkflowInstance,
+        vacancy_start: NaiveDate,
+        target_move_in: NaiveDate,
+        today: NaiveDate,
+    ) -> VacancyInsights {
+        super::generate_insights(self, instance, vacancy_start, target_move_in, today)
+    }
 }
 
 #[derive(Debug)]
@@ -135,12 +104,12 @@ impl TaskSnapshot {
             key: self.key,
             name: self.name,
             stage: self.stage,
-            stage_label: self.stage.label().to_string(),
+            stage_label: self.stage.label(),
             role: self.role,
-            role_label: self.role.label().to_string(),
+            role_label: self.role.label(),
             due_date: self.due_date,
             status: self.status,
-            status_label: self.status.label().to_string(),
+            status_label: self.status.label(),
             completed_on: None,
         }
     }
@@ -161,7 +130,7 @@ impl ComplianceAlert {
             topic: self.topic,
             detail: self.detail,
             severity: self.severity,
-            severity_label: self.severity.label().to_string(),
+            severity_label: self.severity.label(),
         }
     }
 }
